ui | move auth store to auth module; delete config store; code refactor
