--- src/app/classes/users.interface.ts
@@ -1,5 +1,0 @@
-
-export interface IConfig {
-    adminName: string;
-    premissions?: string[];
-}
