+++ src/app/components/header/header.component.ts
@@ -5,13 +5,11 @@ import { select, Store } from '@ngrx/store';
 import { Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 import { AuthModalComponent } from 'src/app/modules/auth/auth-modal/auth-modal.component';
-import { GetConfig } from 'src/app/store/actions/config.actions';
 import { GetUsers } from 'src/app/store/actions/user.actions';
-import { selectConfig } from 'src/app/store/selectors/config.selector';
 import { IError } from 'src/app/store/states/user.state';
 import { IAppState } from 'src/app/store/states/app.state';
 import { IUserAuth } from 'src/app/common/interfaces/auth.interface';
-import { Logout } from 'src/app/store/actions/auth.actions';
+import { Logout } from 'src/app/modules/auth/store/auth.actions';
 
 @Component({
   selector: 'app-header',
@@ -23,30 +21,28 @@ export class HeaderComponent implements OnInit {
   userName: string;
   dropdownVisible = false;
   accessToken: string | null = null;
-  config$ = this.store.pipe(select(selectConfig));
   destroy$ = new Subject<boolean>();
+
   constructor(
     private store: Store<IAppState>,
     private router: Router,
     private dialog: MatDialog
   ) {}
-  isError: IError;
 
   ngOnInit(): void {
-    this.store.dispatch(new GetConfig());
     this.store.dispatch(new GetUsers());
     this.accessToken = localStorage.getItem('accessToken');
     this.isLoggedIn = !!this.accessToken;
     this.userName = localStorage.getItem('username') || '';
-
-    //  this.error$.subscribe(d => this.isError = d);
   }
+
   navigate(event: Event) {
     const target = (event.target as HTMLElement).innerText
       .toLowerCase()
       .replace(' ', '-');
     this.router.navigateByUrl(target);
   }
+
   openAuthModal(page: string) {
     const dialogRef = this.dialog.open(AuthModalComponent, {
       data: {
@@ -54,9 +50,8 @@ export class HeaderComponent implements OnInit {
       },
       autoFocus: false,
     });
-    dialogRef
-      .afterClosed()
-      .pipe(
+
+    dialogRef.afterClosed().pipe(
         takeUntil(this.destroy$),
         filter((action) => !!action)
       )
@@ -71,8 +66,6 @@ export class HeaderComponent implements OnInit {
   }
 
   signOut() {
-    this.isLoggedIn = false;
-    this.userName = '';
     localStorage.clear();
     this.dropdownVisible = !this.dropdownVisible;
     this.store.dispatch(new Logout());
