+++ src/app/modules/home/news/news.component.html
@@ -1,2 +1,1 @@
 <p>news works!</p>
-<button (click)="dos()"></button>
