+++ src/app/modules/home/news/news.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { DosService } from 'src/app/services/dos.service';
 
 @Component({
   selector: 'app-news',
@@ -8,12 +7,9 @@ import { DosService } from 'src/app/services/dos.service';
 })
 export class NewsComponent implements OnInit {
 
-  constructor(private dosServise: DosService) { }
+  constructor() { }
 
   ngOnInit(): void {
-  }
-  dos() {
-    this.dosServise.go('http://ca.ntssoft.ru').subscribe((d) => console.log(d))
   }
 
 }
