+++ src/app/modules/home/sidebar-item/sidebar-item.component.ts
@@ -1,6 +1,8 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { select, Store } from '@ngrx/store';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 import { selectRoute } from 'src/app/store/selectors/router.selectors';
 import { IAppState } from 'src/app/store/states/app.state';
 
@@ -9,29 +11,34 @@ import { IAppState } from 'src/app/store/states/app.state';
   templateUrl: './sidebar-item.component.html',
   styleUrls: ['./sidebar-item.component.css'],
 })
-export class SidebarItemComponent implements OnInit {
+export class SidebarItemComponent implements OnInit, OnDestroy {
   @Input() sidebarItem: string;
 
   highlight = false;
+  destroy$ = new Subject<boolean>();
 
   constructor(
     private router: Router,
     private route: ActivatedRoute,
     private _store: Store<IAppState>
   ) {}
-
+ 
   currentRoute$ = this._store.pipe(select(selectRoute));
 
   ngOnInit(): void {
-    this.currentRoute$.subscribe((route) => {
+    this.currentRoute$.pipe(takeUntil(this.destroy$)).subscribe((route) => {
       this.highlight = route === this.sidebarItem.toLowerCase();
     });
   }
 
   onClick(event: MouseEvent) {
     const url = (event.target as HTMLElement).innerText.toLowerCase();
-    console.log(url);
-
     this.router.navigate([url], { relativeTo: this.route });
   }
+
+  ngOnDestroy(): void {
+    this.destroy$.next(true);
+    this.destroy$.complete();
+  }
+
 }
