--- src/app/store/states/config.state.ts
@@ -1,9 +1,0 @@
-import { IConfig } from "src/app/classes/users.interface";
-
-export interface IConfigState {
-    config: IConfig | null;
-}
-
-export const initialConfigState: IConfigState = {
-    config: null,
-}
